+++ api/v1alpha1/userconfig/service/redis/zz_generated.deepcopy.go
@@ -159,6 +159,16 @@ func (in *RedisUserConfig) DeepCopyInto(out *RedisUserConfig) {
 		*out = make([]string, len(*in))
 		copy(*out, *in)
 	}
+	if in.BackupHour != nil {
+		in, out := &in.BackupHour, &out.BackupHour
+		*out = new(int)
+		**out = **in
+	}
+	if in.BackupMinute != nil {
+		in, out := &in.BackupMinute, &out.BackupMinute
+		*out = new(int)
+		**out = **in
+	}
 	if in.IpFilter != nil {
 		in, out := &in.IpFilter, &out.IpFilter
 		*out = make([]*IpFilter, len(*in))